refactor: move data files to content folder.
