+++ src/components/Galeria.astro
@@ -1,6 +1,6 @@
 ---
 import "photoswipe/style.css"
-import editionsInfo from "@/data/meta-gallery.json"
+import editionsInfo from "@/content/meta/gallery.json"
 import Button from "@/components/Button.astro"
 
 const offset = 10
