+++ src/components/Palmares.astro
@@ -1,6 +1,6 @@
 ---
 import StreamerCard from "@/components/StreamerCard.astro"
-import palmares from "@/data/editions-info.json"
+import palmares from "@/content/editions/info.json"
 
 interface Props {
   edicion: string
