+++ src/content/editions/info.json
