+++ src/content/editions/vote.json
